refactor(worker): remove temporal prefix from module names
Since we're now Temporal-only, simplify the module structure:
- temporal_types.rs → types.rs
- temporal_client.rs → client.rs
- temporal_activities.rs merged into activities.rs
- temporal_runner.rs merged into runner.rs
- run_temporal_worker → run_worker

The worker crate structure is now cleaner with no redundant prefixes.
