+++ crates/everruns-worker/src/session_workflow.rs
@@ -18,7 +18,7 @@ use serde::{Deserialize, Serialize};
 use serde_json::json;
 use tracing::{info, warn};
 
-use crate::temporal_types::*;
+use crate::types::*;
 use crate::workflow_traits::{Workflow, WorkflowInput};
 
 /// Maximum number of tool iterations before forcing completion
